+++ src/commands/poa-whitelist.ts
@@ -1,10 +1,11 @@
 import Vorpal from 'vorpal';
 
-import Node, { Contract } from 'evm-lite-core';
-import utils from 'evm-lite-utils';
+import Node from 'evm-lite-core';
 
 import Session from '../core/Session';
 import Table from '../core/Table';
+
+import { POA } from '../poa';
 
 import Command, { Arguments, TxOptions } from '../core/TxCommand';
 
@@ -41,66 +42,6 @@ export default (evmlc: Vorpal, session: Session) => {
 };
 
 class POAWhitelistCommand extends Command<Args> {
-	public async getWhitelist() {
-		const poa = await this.node!.getPOA();
-
-		this.log.info('POA', poa.address);
-
-		const contract = Contract.load(JSON.parse(poa.abi), poa.address);
-
-		const transaction = contract.methods.getWhiteListCount({
-			gas: this.args.options.gas,
-			gasPrice: Number(this.args.options.gasPrice)
-		});
-
-		const countRes: any = await this.node!.callTx(transaction);
-		const count = countRes.toNumber();
-		this.debug(`Whitelist count -> ${count}`);
-
-		if (!count) {
-			return [];
-		}
-
-		// entries
-		const entries: WhitelistEntry[] = [];
-
-		for (const i of Array.from(Array(count).keys())) {
-			const entry: WhitelistEntry = {
-				address: '',
-				moniker: ''
-			};
-
-			const addressTx = contract.methods.getWhiteListAddressFromIdx(
-				{
-					gas: this.args.options.gas,
-					gasPrice: Number(this.args.options.gasPrice)
-				},
-				i
-			);
-
-			entry.address = await this.node!.callTx(addressTx);
-
-			const monikerTx = contract.methods.getMoniker(
-				{
-					gas: this.args.options.gas,
-					gasPrice: Number(this.args.options.gasPrice)
-				},
-				entry.address
-			);
-
-			const hex = await this.node!.callTx<string>(monikerTx);
-			entry.moniker = utils.hexToString(hex);
-
-			this.debug(
-				`Adding whitelist entry -> ${entry.moniker} (${entry.address})`
-			);
-
-			entries.push(entry);
-		}
-
-		return entries;
-	}
-
 	public async init(): Promise<boolean> {
 		this.constant = true;
 
@@ -132,7 +73,10 @@ class POAWhitelistCommand extends Command<Args> {
 			`${this.args.options.host}:${this.args.options.port}/poa`
 		);
 
-		const entries = (await this.getWhitelist()) as WhitelistEntry[];
+		const poa = new POA(this.args.options.host, this.args.options.port);
+		await poa.init();
+
+		const entries = await poa.whitelist();
 		const table = new Table(['Moniker', 'Address']);
 
 		if (!entries.length) {