+++ src/evmlc.ts
@@ -7,8 +7,8 @@ import accountsCreate from './commands/accounts-create';
 import accountsGenerate from './commands/accounts-generate';
 import accountsGet from './commands/accounts-get';
 import accountsImport from './commands/accounts-import';
-import accountsList from './commands/accounts-list';
 import accountsInspect from './commands/accounts-inspect';
+import accountsList from './commands/accounts-list';
 import accountsUpdate from './commands/accounts-update';
 
 import configSet from './commands/config-set';