+++ src/poa/index.ts
@@ -1,0 +1,8 @@
+export {
+	default as POA,
+	NomineeEntry,
+	WhitelistEntry,
+	EvicteeEntry
+} from './Contract';
+export { default as Logs } from './Logs';
+export * from './Events';